requirements.txt
wheel
matplotlib~=3.6.1
numpy~=1.23.3
pandas~=1.5.0
seaborn~=0.12.0
jupyterlab
pyreadstat~=1.1.9
statsmodels~=0.13.2
sklearn~=0.0
stata_setup
stata_kernel
lxml
dbf~=0.99.2
yattag~=1.14.0
pyarrow
python-docx
openpyxl
markdown
omoment~=0.1.2
reportree~=0.0.3
requests~=2.28.1
scipy~=1.9.1
scikit-learn~=1.1.2
openai~=0.25.0
pytdigest
ipython
